Remove webviews support (#226)
