--- connect/res/layout/com_telenor_connect_activity_layout.xml
@@ -1,4 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-             android:id="@+id/com_telenor_connect_fragment_container"
-             android:layout_width="match_parent"
-             android:layout_height="match_parent" />
