--- connect/res/layout/com_telenor_connect_error_view.xml
@@ -1,132 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/com_telenor_connect_darker_gray"
-    >
-    <ScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:fillViewport="true"
-        >
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            android:padding="16dp"
-            >
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:orientation="vertical"
-                android:gravity="center_horizontal"
-                android:layout_weight="1"
-                >
-                <TextView
-                    android:id="@+id/com_telenor_connect_error_view_communication_error"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="100dp"
-                    android:gravity="center"
-                    android:textColor="@color/com_telenor_connect_light_color"
-                    android:paddingLeft="50dp"
-                    android:paddingRight="50dp"
-                    android:textSize="16sp"
-                    />
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:gravity="center"
-                    android:paddingLeft="50dp"
-                    android:paddingRight="50dp"
-                    android:text="@string/com_telenor_connect_error_text2"
-                    android:textColor="@color/com_telenor_connect_light_color"
-                    android:layout_marginTop="16sp"
-                    android:textSize="16sp"
-                    />
-                <TextView
-                    android:id="@+id/com_telenor_connect_error_view_text"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:paddingLeft="50dp"
-                    android:paddingRight="50dp"
-                    android:textColor="@color/com_telenor_connect_light_gray"
-                    android:layout_marginTop="20dp"
-                    android:textSize="12sp"
-                    />
-                <Button
-                    android:id="@+id/com_telenor_connect_error_view_show_more_toggle"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/com_telenor_connect_show_details"
-                    style="@style/Widget.AppCompat.Button.Borderless"
-                    android:textColor="@color/com_telenor_connect_light_color"
-                    android:textSize="12sp"
-                    />
-                <TextView
-                    android:id="@+id/com_telenor_connect_error_view_details"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:textColor="@color/com_telenor_connect_light_gray"
-                    android:textSize="12sp"
-                    android:gravity="center"
-                    android:visibility="gone"
-                    />
-            </LinearLayout>
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical"
-                android:gravity="center"
-                >
-                <Button
-                    android:id="@+id/com_telenor_connect_error_view_try_again"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="@string/com_telenor_connect_try_again"
-                    android:theme="@style/Theme.AppCompat.Light"
-                    android:paddingTop="16sp"
-                    android:paddingBottom="16sp"
-                    android:textSize="16sp"
-                    android:layout_marginTop="10sp"
-                    />
-                <Button
-                    android:id="@+id/com_telenor_connect_error_view_network_settings"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    style="@style/Widget.AppCompat.Button.Borderless"
-                    android:text="@string/com_telenor_connect_network_settings"
-                    android:layout_marginTop="8sp"
-                    android:textSize="16sp"
-                    android:paddingTop="16sp"
-                    android:paddingBottom="16sp"
-                    android:layout_marginBottom="40dp"
-                    android:textColor="@color/com_telenor_connect_light_color"
-                    />
-                <ImageView
-                    android:visibility="gone"
-                    android:id="@+id/com_telenor_connect_error_view_footer"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:src="@drawable/connect_by_telenordigital"
-                    android:layout_marginBottom="10dp"
-                    />
-            </LinearLayout>
-        </LinearLayout>
-    </ScrollView>
-    <FrameLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id="@+id/com_telenor_connect_error_view_loading"
-        android:background="@color/com_telenor_connect_darker_gray_transparent"
-        android:visibility="invisible"
-        >
-        <ProgressBar
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:theme="@style/ConnectTheme.Dark"
-            android:layout_gravity="center"
-            />
-    </FrameLayout>
-
-</FrameLayout>
