+++ connect/src/com/telenor/connect/BrowserType.java
@@ -1,7 +1,7 @@
 package com.telenor.connect;
 
 public enum BrowserType {
-    WEB_VIEW("web-view"),
+    EXTERNAL_BROWSER("external-browser"),
     CHROME_CUSTOM_TAB("chrome-custom-tab");
 
     private String versionString;
