+++ connect/src/com/telenor/connect/ui/ConnectCustomTabLoginButton.java
@@ -20,16 +20,21 @@ import android.view.View;
 
 import com.telenor.connect.BrowserType;
 import com.telenor.connect.ConnectSdk;
+import com.telenor.connect.R;
+import com.telenor.connect.headerenrichment.ShowLoadingCallback;
+import com.telenor.connect.id.Claims;
+import com.telenor.connect.id.IdProvider;
+import com.telenor.connect.utils.ClaimsParameterFormatter;
 import com.telenor.connect.utils.ConnectUrlHelper;
 import com.telenor.connect.utils.CustomTabsHelper;
 
 import java.lang.ref.WeakReference;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
 
-/**
- * Uses custom tabs, unless not available. If not available falls back to ConnectWebViewLoginButton
- * logic.
- */
-class ConnectCustomTabLoginButton extends ConnectWebViewLoginButton {
+class ConnectCustomTabLoginButton extends ConnectButton implements AuthenticationButton {
 
     private static final Uri PRE_FETCH_URL
             = Uri.parse(
@@ -48,17 +53,20 @@ class ConnectCustomTabLoginButton extends ConnectWebViewLoginButton {
     private boolean serviceBound = false;
     private BrowserType browserType;
     private CustomTabsSession session;
+    private ArrayList<String> acrValues;
+    private Map<String, String> loginParameters;
+    private ArrayList<String> loginScopeTokens;
+    private int requestCode = 0xa987;
+    private Claims claims;
+    private ShowLoadingCallback showLoadingCallback;
 
     public ConnectCustomTabLoginButton(final Context context, AttributeSet attributeSet) {
         super(context, attributeSet);
+        IdProvider brand = ConnectSdk.getIdProvider();
+        setText(getResources().getString(R.string.com_telenor_connect_login_button_text, getResources().getString(brand.getNameKey())));
     }
 
-    @Override
     protected void authenticate() {
-        if (!customTabsSupported) {
-            super.authenticate();
-            return;
-        }
         ConnectSdk.authenticate(
                 session,
                 getParameters(),
@@ -78,6 +86,32 @@ class ConnectCustomTabLoginButton extends ConnectWebViewLoginButton {
         return context.getPackageName().equals(componentName.getPackageName());
     }
 
+    @NonNull
+    protected Map<String, String> getParameters() {
+        Map<String, String> parameters = new HashMap<>();
+        if (getAcrValues() != null && !getAcrValues().isEmpty()) {
+            parameters.put("acr_values", TextUtils.join(" ", getAcrValues()));
+        }
+
+        if (getLoginScopeTokens() != null && !getLoginScopeTokens().isEmpty()) {
+            parameters.put("scope", TextUtils.join(" ", getLoginScopeTokens()));
+        }
+
+        addClaims(parameters);
+
+        if (getLoginParameters() != null && !getLoginParameters().isEmpty()) {
+            parameters.putAll(getLoginParameters());
+        }
+
+        return parameters;
+    }
+
+    private void addClaims(Map<String, String> parameters) {
+        if (getClaims() != null && getClaims().getClaimsAsSet() != null) {
+            parameters.put("claims", ClaimsParameterFormatter.asJson(getClaims()));
+        }
+    }
+
     @Override
     protected void onAttachedToWindow() {
         super.onAttachedToWindow();
@@ -93,7 +127,7 @@ class ConnectCustomTabLoginButton extends ConnectWebViewLoginButton {
         serviceBound = CustomTabsClient.bindCustomTabsService(getContext(), packageNameToUse, connection);
         boolean correctIntentFilter = contextIntentFilterMatchesRedirectUri(getContext());
         customTabsSupported = serviceBound && correctIntentFilter;
-        browserType = customTabsSupported ? BrowserType.CHROME_CUSTOM_TAB : BrowserType.WEB_VIEW;
+        browserType = customTabsSupported ? BrowserType.CHROME_CUSTOM_TAB : BrowserType.EXTERNAL_BROWSER;
     }
 
     @Override
@@ -119,6 +153,63 @@ class ConnectCustomTabLoginButton extends ConnectWebViewLoginButton {
 
     private void setSession(CustomTabsSession session) {
         this.session = session;
+    }
+
+    public ShowLoadingCallback getShowLoadingCallback() {
+        return showLoadingCallback;
+    }
+
+    public void setShowLoadingCallback(ShowLoadingCallback showLoadingCallback) {
+        this.showLoadingCallback = showLoadingCallback;
+    }
+
+
+    public ArrayList<String> getAcrValues() {
+        return acrValues;
+    }
+
+    public Map<String, String> getLoginParameters() {
+        return loginParameters;
+    }
+
+    public ArrayList<String> getLoginScopeTokens() {
+        return loginScopeTokens;
+    }
+
+    public int getRequestCode() {
+        return requestCode;
+    }
+
+    public Claims getClaims() {
+        return claims;
+    }
+
+    public void setAcrValues(String... acrValues) {
+        this.acrValues = new ArrayList<>(Arrays.asList(acrValues));
+    }
+
+    public void setAcrValues(ArrayList<String> acrValues) {
+        this.acrValues = acrValues;
+    }
+
+    public void setLoginScopeTokens(String... scopeTokens) {
+        loginScopeTokens = new ArrayList<>(Arrays.asList(scopeTokens));
+    }
+
+    public void setLoginScopeTokens(ArrayList<String> scopeTokens) {
+        loginScopeTokens = scopeTokens;
+    }
+
+    public void setExtraLoginParameters(Map<String, String> parameters) {
+        loginParameters = parameters;
+    }
+
+    public void setRequestCode(int requestCode) {
+        this.requestCode = requestCode;
+    }
+
+    public void setClaims(Claims claims) {
+        this.claims = claims;
     }
 
     private static class WeakReferenceCustomTabsServiceConnection extends CustomTabsServiceConnection {
